% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_to_disc_M.R
\name{cont_to_disc_M}
\alias{cont_to_disc_M}
\title{Convert data to M statistic}
\usage{
cont_to_disc_M(ICdata, OCdata, mu.p = mean(ICdata))
}
\arguments{
\item{ICdata}{The IC data}

\item{OCdata}{The OC data}

\item{mu.p}{Population mean, when the population mean is unknown, it will be estimated by the sample mean.}
}
\value{
M0 is M statistic for IC data.

M1 is M statistic for OC data.

p0 is the process proportion for IC data.

p1 is the process proportion for OC data.

n is the number of the sample size.
}
\description{
Convert continuity data to discrete data with M statistic
}
\examples{
IC = matrix(rnorm(100,0,1),ncol = 10,byrow = TRUE)
OC = matrix(rnorm(100,2,1),ncol = 10,byrow = TRUE)
cont_to_disc_M(IC,OC)
}
\references{
Yang, S. F., & Arnold, B. C. (2014). A simple approach for monitoring business service time variation. The Scientific World Journal, 2014.
}
