% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewma.R
\name{ewma}
\alias{ewma}
\title{EWMA charting statistic of input data}
\usage{
ewma(data, lambda, EWMA0)
}
\arguments{
\item{data}{The data what to calculate EWMA charting statistic}

\item{lambda}{The EWMA smooth constant}

\item{EWMA0}{The starting point of EWMA charting statistic}
}
\value{
The EWMA charting statistic of data.
}
\description{
EWMA charting statistic of input data
}
\examples{
x = rnorm(20,0,1)
ewma(x,0.05,0)
}
