% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWMA_p_chart_one_UCL.R
\name{EWMA_p_chart_one_UCL}
\alias{EWMA_p_chart_one_UCL}
\title{Plot one-sided upper EWMA-p control chart}
\usage{
EWMA_p_chart_one_UCL(
  ICdata,
  OCdata,
  lambda,
  n,
  pi1 = 1,
  pi2 = pi1,
  ARL0 = 200,
  M = 500,
  error = 10
)
}
\arguments{
\item{ICdata}{The IC data for attributes}

\item{OCdata}{The OC data for attributes}

\item{lambda}{The EWMA smooth constant}

\item{n}{The number of sample size}

\item{pi1}{The correctness of true}

\item{pi2}{The correctness of false}

\item{ARL0}{The ARL of in-control process}

\item{M}{The simulation times of Monte Carlo method}

\item{error}{The maximum error what we can allow between hat_ARL0 and ARL0}
}
\value{
The first chart is EWMA-p chart for IC data, and the second chart is EWMA-p chart for OC data
}
\description{
Plot one-sided upper EWMA-p control chart  for input IC and OC data, the input need to be number of defectives.
}
\examples{
library(qcr)
data = orangejuice
IC = data[31:54,1]
OC = data[1:30,1]
EWMA_p_chart_one_UCL(IC,OC,0.05,50,1,1)
}
