% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWMA_p_one_LCL.R
\name{EWMA_p_one_LCL}
\alias{EWMA_p_one_LCL}
\title{The one-sided lower control limit for EWMA-p chart}
\usage{
EWMA_p_one_LCL(
  p,
  lambda,
  n,
  pi1 = 1,
  pi2 = pi1,
  ARL0 = 200,
  M = 500,
  error = 10
)
}
\arguments{
\item{p}{The in-control probability}

\item{lambda}{The EWMA smooth constant}

\item{n}{The number of sample size}

\item{pi1}{The correctness of true}

\item{pi2}{The correctness of false}

\item{ARL0}{The ARL of in-control process}

\item{M}{The simulation times of Monte Carlo method}

\item{error}{The maximum error what we can allow between hat_ARL0 and ARL0}
}
\value{
L2 is a constant used to specify the width of the lower control limits.

hat_ARL0 is the simulated result for average run length with L2.

hat_MRL0 is the simulated result for median of run length with L2.

hat_SDRL0 is the simulated result for standard deviation of run length with L2.

LCL is the limiting values of lower control limits with L2.
}
\description{
This function can calculate the one-sided lower control limit for EWMA-p chart, and adjusted EWMA-p chart.
The parameters pi1 and pi2 are the correct rate of measured valued.
If we do not consider about measurement error, the value of pi1 and pi2 are 1.
}
\examples{
EWMA_p_one_LCL(0.2,0.05,5,1,1)
}
