% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_data_generate.R
\name{ME_data_generate}
\alias{ME_data_generate}
\title{Generate the discrete random variable with measurement error}
\usage{
ME_data_generate(p, n, m, pi1, pi2 = pi1)
}
\arguments{
\item{p}{A probability of the unobserved defectives.}

\item{n}{A number of sample size in the data.}

\item{m}{A number of observation in the data.}

\item{pi1}{The proportion that the observed defectives are the same as unobserved ones.}

\item{pi2}{The proportion that the observed non-defectives are the same as unobserved ones.}
}
\value{
\code{real_data}\eqn{\hspace{1.2cm}} The generated data without measurement error.

\code{obs_data}\eqn{\hspace{1.4cm}} The generated data with measurement error.

\code{n}\eqn{\hspace{2.6cm}} A sample size in the generated data.
}
\description{
Generate the discrete random variable with measurement error.
}
\examples{
ME_data_generate(0.7,50,50,0.95)
}
