% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_to_disc_V.R
\name{cont_to_disc_V}
\alias{cont_to_disc_V}
\title{Convert data to V statistic}
\usage{
cont_to_disc_V(ICdata, OCdata, var.p = NULL)
}
\arguments{
\item{ICdata}{The IC data}

\item{OCdata}{The OC data}

\item{var.p}{Population variance, when the population variance is unknown, it will be estimated by the square of sample standard deviation.}
}
\value{
V0 is V statistic for IC data.

V1 is V statistic for OC data.

p0 is the process proportion for IC data.

p1 is the process proportion for OC data.

n is the number of the sample size.
}
\description{
Convert continuity data to discrete data with V statistic
}
\examples{
IC = matrix(rnorm(100,0,1),ncol = 10,byrow = TRUE)
OC = matrix(rnorm(100,0,2),ncol = 10,byrow = TRUE)
cont_to_disc_V(IC,OC)
}
\references{
Yang, S. F., & Arnold, B. C. (2014). A simple approach for monitoring business service time variation. The Scientific World Journal, 2014.
}
