% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_data_generate.R
\name{ME_data_generate}
\alias{ME_data_generate}
\title{Generate the discrete data with measurement error}
\usage{
ME_data_generate(p, n, m, pi1, pi2 = pi1)
}
\arguments{
\item{p}{The probability}

\item{n}{The number of sample size}

\item{m}{The number of observation}

\item{pi1}{The correctness of true}

\item{pi2}{The correctness of false}
}
\value{
real_data is the process proportion for IC data.

obs_data is the process proportion for OC data.

n is the number of the sample size.
}
\description{
Generate the discrete data with measurement error
}
\examples{
ME_data_generate(0.7,50,50,0.95)
}
